+++ src/services/cls-namespace.service.ts
@@ -4,10 +4,14 @@ import { Service } from '../decorators/services/service.decorator';
 
 @Service()
 export class ClsNamespaceService {
-    constructor(private meta: ServerMetadata) {
-        let _name = `miter-cls-(${meta.name || ''})-${ClsNamespaceService.namespaceNum++}`;
+    constructor(
+        private meta: ServerMetadata
+    ) {
+        let _name = `miter-cls-(${this.meta.name || ''})-${ClsNamespaceService.namespaceNum++}`;
         this._namespace = createNamespace(_name);
     }
+    
+    async start() { }
     
     private static namespaceNum: number = 0;
     private _namespace: Namespace;